fix: race between VolumeConfigController and UserVolumeConfigController
This commit just copy/paste merges the VolumeConfig and UserVolumeConfig
controllers, which addresses the race between system volume and user
volume config creation.
